api/requirements.txt
django
djangorestframework
django-nested-admin
django-cors-headers
